+++ internal/backend/front-event.go
@@ -4,4 +4,5 @@ const (
 	FEExitDirty        = "exitDirty"
 	FEShowError        = "showError"
 	FEUpgradeAvailable = "upgradeAvailable"
+	FEOpenFiles        = "openFiles"
 )