+++ main.go
@@ -1,13 +1,17 @@
 package main
 
 import (
+	"context"
 	"embed"
 	"fmt"
 	"os"
 
+	"github.com/pkg/errors"
 	"github.com/vovan-ve/sple-desktop/internal/backend"
 	"github.com/vovan-ve/sple-desktop/internal/config"
+	"github.com/vovan-ve/sple-desktop/internal/files"
 	"github.com/vovan-ve/sple-desktop/internal/logging"
+	"github.com/vovan-ve/sple-desktop/internal/singleton"
 	"github.com/wailsapp/wails/v2"
 	"github.com/wailsapp/wails/v2/pkg/logger"
 	"github.com/wailsapp/wails/v2/pkg/options"
@@ -15,6 +19,8 @@ import (
 	"github.com/wailsapp/wails/v2/pkg/options/linux"
 	"github.com/wailsapp/wails/v2/pkg/options/windows"
 )
+
+const appTitle = "SpLE"
 
 //go:embed all:frontend/build-wails
 var assets embed.FS
@@ -31,18 +37,30 @@ func main() {
 			lg.Fatal(fmt.Sprintf("PANIC recovery: %+v\n", r))
 		} else {
 			println("PANIC recovery:", r)
+			os.Exit(1)
 		}
-		os.Exit(1)
 	}()
-
-	// Create an instance of the app structure
-	app := NewApp()
 
 	lg = logging.GetLogger()
 
+	argFiles, argsErr := files.NormalizeArgs(os.Args[1:])
+	if done, err := sendSingletonArgs(argFiles, lg); err != nil {
+		lg.Fatal(fmt.Sprintf("singleton: %v", err))
+	} else if done {
+		lg.Info("The primary instance will do the job. Exiting.")
+		os.Exit(0)
+	}
+
+	// Create an instance of the app structure
+	app := NewApp(&AppOptions{
+		Logger:           lg,
+		StartupOpenFiles: argFiles,
+		StartupError:     argsErr,
+	})
+
 	// Create application with options
 	err := wails.Run(&options.App{
-		Title:                    "SpLE",
+		Title:                    appTitle,
 		Width:                    1024,
 		Height:                   768,
 		EnableDefaultContextMenu: true,
@@ -75,7 +93,7 @@ func main() {
 			WebviewUserDataPath:  config.GetConfigsDir(),
 		},
 		Linux: &linux.Options{
-			ProgramName: "SpLE",
+			ProgramName: appTitle,
 		},
 		//Mac: &mac.Options{},
 		Debug: options.Debug{
@@ -83,10 +101,19 @@ func main() {
 		},
 	})
 	if err != nil {
-		if lg != nil {
-			lg.Error(err.Error())
-		} else {
-			println("Error:", err)
-		}
+		lg.Fatal(err.Error())
 	}
+}
+
+func sendSingletonArgs(args []string, l logger.Logger) (done bool, err error) {
+	l.Debug(fmt.Sprintf("IPC args: %#v", args))
+	messages, err := files.ArgsToIpc(args)
+	if err != nil {
+		return false, errors.Wrap(err, "ArgsToIpc")
+	}
+	done, err = singleton.TrySend(context.TODO(), l, messages)
+	if err != nil {
+		return false, errors.Wrap(err, "TrySend")
+	}
+	return done, nil
 }