+++ CHANGELOG.md
@@ -5,6 +5,12 @@
 - Add: Textual editing for demo keystrokes.
 - Fix: SP file format loss demo signature on file read.
 
+### Desktop
+
+- Add: Prevent multiple instances to ran in the same time.
+- Add: "Open file with" like whatever editors do.
+- Fix: Same file now cannot be opened multiple times in same instance.
+
 ## 0.16.1 (2023-09-18)
 
 ### Desktop